Extract data access logics to repository layer from service layer, which means repository layer could be seen as a bridge between service layer and mapper layer
