+++ src/main/java/com/bryan/system/model/vo/UserExportVO.java
@@ -25,6 +25,7 @@ import java.util.Map;
 @Data
 @Builder
 public class UserExportVO {
+
     @ExcelProperty("用户ID")
     @ColumnWidth(10)
     private Long id;
