+++ src/main/java/com/bryan/system/repository/UserProfileRepository.java
@@ -1,0 +1,18 @@
+package com.bryan.system.repository;
+
+import com.bryan.system.mapper.UserProfileMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+/**
+ * UserProfileRepository
+ *
+ * @author Bryan Long
+ */
+@Repository
+@RequiredArgsConstructor
+public class UserProfileRepository {
+
+    private final UserProfileMapper userProfileMapper;
+
+}
